Remove temporary /init and poweroff now that we have systemd.
The stage4 kernel is still compiled with init=/init.  That's the only
reasonable thing to do considering (a) we are forced to compile the
command line into the kernel image and (b) we wish to replace init in
some circumstances (eg in the autobuilder).

Now we make /init be a symlink to systemd.  In cases like the
autobuilder where we want to use our own init, the /init symlink can
be replaced by a shell script, but the normal case now is to boot into
systemd and run the usual systemd units.

The poweroff command is removed entirely, since systemd supplies a
poweroff command.